Package structure refactor.
