+++ app/src/main/java/com/baldev/eventify/domain/actions/groups/DefaultAddUsersToGroupAction.java
@@ -1,4 +1,4 @@
-package com.baldev.eventify.domain.actions;
+package com.baldev.eventify.domain.actions.groups;
 
 import com.baldev.eventify.domain.entities.Group;
 import com.baldev.eventify.domain.entities.User;
