+++ app/src/main/java/com/baldev/eventify/domain/actions/users/DefaultGetMyUserAction.java
@@ -1,4 +1,4 @@
-package com.baldev.eventify.domain.actions;
+package com.baldev.eventify.domain.actions.users;
 
 
 import com.baldev.eventify.domain.entities.User;
