+++ app/src/main/java/com/baldev/eventify/domain/actions/users/SaveUserCallback.java
@@ -1,4 +1,4 @@
-package com.baldev.eventify.domain.actions;
+package com.baldev.eventify.domain.actions.users;
 
 public interface SaveUserCallback {
 	void onUserSaved();
