+++ app/src/main/java/com/baldev/eventify/domain/repositories/UsersRepository.java
@@ -1,7 +1,7 @@
 package com.baldev.eventify.domain.repositories;
 
 
-import com.baldev.eventify.domain.actions.SaveUserCallback;
+import com.baldev.eventify.domain.actions.users.SaveUserCallback;
 import com.baldev.eventify.domain.entities.User;
 
 public interface UsersRepository {
