+++ app/src/main/java/com/baldev/eventify/domain/services/SaveUserService.java
@@ -1,7 +1,7 @@
 package com.baldev.eventify.domain.services;
 
 
-import com.baldev.eventify.domain.actions.SaveUserCallback;
+import com.baldev.eventify.domain.actions.users.SaveUserCallback;
 import com.baldev.eventify.domain.entities.User;
 import com.baldev.eventify.domain.repositories.UsersRepository;
