+++ app/src/main/java/com/baldev/eventify/presentation/creategroup/CreateGroupPresenter.java
@@ -1,7 +1,7 @@
 package com.baldev.eventify.presentation.creategroup;
 
-import com.baldev.eventify.domain.actions.CreateGroupAction;
-import com.baldev.eventify.domain.actions.GetMyUserAction;
+import com.baldev.eventify.domain.actions.groups.CreateGroupAction;
+import com.baldev.eventify.domain.actions.users.GetMyUserAction;
 import com.baldev.eventify.domain.entities.User;
 import com.baldev.eventify.presentation.creategroup.CreateGroupContract.Presenter;
 import com.baldev.eventify.presentation.creategroup.CreateGroupContract.View;
