+++ app/src/test/java/com/baldev/eventify/domain/actions/GetMyUserActionTest.java
@@ -1,5 +1,7 @@
 package com.baldev.eventify.domain.actions;
 
+import com.baldev.eventify.domain.actions.users.DefaultGetMyUserAction;
+import com.baldev.eventify.domain.actions.users.GetMyUserAction;
 import com.baldev.eventify.domain.entities.User;
 import com.baldev.eventify.domain.repositories.UsersRepository;
