+++ app/src/test/java/com/baldev/eventify/domain/actions/SaveUserActionTest.java
@@ -1,5 +1,7 @@
 package com.baldev.eventify.domain.actions;
 
+import com.baldev.eventify.domain.actions.users.DefaultSaveUserAction;
+import com.baldev.eventify.domain.actions.users.SaveUserCallback;
 import com.baldev.eventify.domain.entities.User;
 import com.baldev.eventify.domain.services.SaveUserService;
