+++ app/src/test/java/com/baldev/eventify/presentation/creategroup/CreateGroupPresenterTest.java
@@ -1,7 +1,7 @@
 package com.baldev.eventify.presentation.creategroup;
 
-import com.baldev.eventify.domain.actions.CreateGroupAction;
-import com.baldev.eventify.domain.actions.GetMyUserAction;
+import com.baldev.eventify.domain.actions.groups.CreateGroupAction;
+import com.baldev.eventify.domain.actions.users.GetMyUserAction;
 import com.baldev.eventify.domain.entities.User;
 
 import org.junit.Before;
